+++ api/services/benchmark/store_createbenchmark.go
@@ -3,6 +3,7 @@ package benchmark
 import (
 	"context"
 	"errors"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
@@ -23,11 +24,14 @@ func (s *PostgresBenchmarkStore) CreateBenchmark(ctx context.Context, b *types.B
 
 	row := s.db.QueryRow(
 		c,
-		`INSERT INTO benchmarks
-		(name, description, asset_allocation, std_dev_pct, real_return_pct, drawdown_yrs, is_deprecated, user_id, rec_rebalance_threshold_pct)
-		VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)
-		RETURNING *`,
-		b.Name, b.Description, b.Asset_allocation, b.Std_dev_pct, b.Real_return_pct, b.Drawdown_yrs, b.Is_deprecated, b.User_id, rec_rebalance_threshold_pct,
+		fmt.Sprintf(`
+			insert into benchmarks
+				(name, description, asset_allocation, std_dev_pct, real_return_pct, drawdown_yrs, rec_rebalance_threshold_pct, is_deprecated, user_id)
+				VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)
+			returning
+				%s
+		`, benchmarkColumns),
+		b.Name, b.Description, b.Asset_allocation, b.Std_dev_pct, b.Real_return_pct, b.Drawdown_yrs, rec_rebalance_threshold_pct, b.Is_deprecated, b.User_id,
 	)
 
 	benchmark, err := s.parseRowIntoBenchmark(row)
