+++ api/services/benchmark/store_deletebenchmark.go
@@ -2,6 +2,7 @@ package benchmark
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
@@ -13,15 +14,15 @@ func (s *PostgresBenchmarkStore) DeleteBenchmark(ctx context.Context, userId, be
 
 	row := s.db.QueryRow(
 		c,
-		`
+		fmt.Sprintf(`
 			delete from
 				benchmarks
 			where
 				benchmark_id = $1
 				and user_id = $2
 			returning
-				*
-		`,
+				%s
+		`, benchmarkColumns),
 		benchmarkId,
 		userId,
 	)
