+++ api/services/benchmark/store_getbenchmarkbyid.go
@@ -2,10 +2,10 @@ package benchmark
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
-	"github.com/jackc/pgx/v5"
 )
 
 func (s *PostgresBenchmarkStore) GetBenchmarkById(ctx context.Context, userId, benchmarkId int) (types.Benchmark, error) {
@@ -14,15 +14,15 @@ func (s *PostgresBenchmarkStore) GetBenchmarkById(ctx context.Context, userId, b
 
 	row := s.db.QueryRow(
 		c,
-		`
+		fmt.Sprintf(`
 			select
-				* 
+				%s 
 			from 
 				benchmarks
 			where 
 				benchmark_id = $1
 				and user_id = $2
-		`,
+		`, benchmarkColumns),
 		benchmarkId,
 		userId,
 	)
@@ -33,27 +33,4 @@ func (s *PostgresBenchmarkStore) GetBenchmarkById(ctx context.Context, userId, b
 		return types.Benchmark{}, err
 	}
 	return benchmark, nil
-}
-
-func (s *PostgresBenchmarkStore) parseRowIntoBenchmark(row pgx.Row) (types.Benchmark, error) {
-	var b types.Benchmark
-	err := row.Scan(
-		&b.Benchmark_id,
-		&b.Name,
-		&b.Description,
-		&b.Asset_allocation,
-		&b.Std_dev_pct,
-		&b.Real_return_pct,
-		&b.Drawdown_yrs,
-		&b.Is_deprecated,
-		&b.User_id,
-		&b.Created_at,
-		&b.Updated_at,
-		&b.Rec_rebalance_threshold_pct,
-	)
-
-	if err != nil {
-		return types.Benchmark{}, err
-	}
-	return b, nil
 }
