+++ api/stores.txt
@@ -1,0 +1,43 @@
+./api/services/snapshot/store_groupbymaturationdate.go
+./api/services/snapshot/store_getsnapshotbydate.go
+./api/services/snapshot/store_getsnapshotbyid.go
+./api/services/snapshot/store_getsnapshottotal.go
+./api/services/snapshot/store.go
+./api/services/snapshot/store_deletesnapshot.go
+./api/services/snapshot/store_createsnapshot.go
+./api/services/snapshot/store_groupbyholding.go
+./api/services/snapshot/store_groupbyliquidity.go
+./api/services/snapshot/store_updatesnapshot.go
+./api/services/snapshot/store_refreshweighteder.go
+./api/services/snapshot/store_getsnapshots.go
+./api/services/snapshot/store_refreshtotal.go
+./api/services/snapshot/store_groupbyaccount.go
+./api/services/user/store_updateuserloggedin.go
+./api/services/user/store_getuserbyid.go
+./api/services/user/store.go
+./api/services/user/store_deleteuser.go
+./api/services/user/store_getuserbyemail.go
+./api/services/user/store_updateverification.go
+./api/services/user/store_createuser.go
+./api/services/user/store_getusers.go
+./api/services/account/store_deleteaccount.go
+./api/services/account/store.go
+./api/services/account/store_getaccounts.go
+./api/services/account/store_getaccountbyid.go
+./api/services/account/store_updateaccount.go
+./api/services/account/store_getaccountbyname.go
+./api/services/account/store_createaccount.go
+./api/services/snapshotvalue/store.go
+./api/services/snapshotvalue/store_updatesnapshotvalue.go
+./api/services/snapshotvalue/store_deletesnapshotvalue.go
+./api/services/snapshotvalue/store_createsnapshotvalue.go
+./api/services/snapshotvalue/store_getsnapshotvalue.go
+./api/services/snapshotvalue/store_getsnapshotvalues.go
+./api/services/holding/store_updateholding.go
+./api/services/holding/store.go
+./api/services/holding/store_getholdingbyname.go
+./api/services/holding/store_createholding.go
+./api/services/holding/store_getholdings.go
+./api/services/holding/store_getholdingbyticker.go
+./api/services/holding/store_deleteholding.go
+./api/services/holding/store_getholdingbyid.go
